pyproject.toml
[project]
name = "mtnbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy>=3.0.4",
    "fastapi[standard]>=0.115.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
    "langwatch>=0.7.2",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.40.0",
    "uvicorn>=0.34.0",
]
